+++ packages/docs-page/index.tsx
@@ -30,6 +30,7 @@ interface DocsPageWrapperProps {
   baseRoute: string
   githubFileUrl: string
   product: { name: string; slug: string }
+  /** @deprecated */
   showEditPage: boolean
   showVersionSelect: boolean
   versions: { name: string; label: string }[]
@@ -132,7 +133,7 @@ const DocsPageWrapper: FunctionComponent<DocsPageWrapperProps> = ({
         </div>
       </div>
       {/* if desired, show an "edit this page" link on the bottom right, linking to github */}
-      {showEditPage && (
+      {showEditPage && githubFileUrl && (
         <div className={s.editThisPage}>
           <a href={githubFileUrl}>
             <img src={require('./img/github-logo.svg')} alt="github logo" />
@@ -147,9 +148,9 @@ const DocsPageWrapper: FunctionComponent<DocsPageWrapperProps> = ({
 export interface DocsPageProps {
   product: { name: string; slug: string }
   baseRoute: string
-  showEditPage: boolean
-  showVersionSelect: boolean
-  additionalComponents: MDXProviderComponentsProp
+  showEditPage?: boolean
+  showVersionSelect?: boolean
+  additionalComponents?: MDXProviderComponentsProp
   staticProps: {
     mdxSource: MDXRemoteSerializeResult
     frontMatter: {