+++ packages/docs-page/server/resolve-nav-data.test.ts
@@ -1,0 +1,5 @@
+import { resolveNavData } from './resolve-nav-data'
+
+describe('resolveNavData', () => {
+  it.todo('should resolve nav data')
+})