+++ packages/version-select/index.test.tsx
@@ -13,9 +13,9 @@ const useRouterMock = mocked(useRouter)
 const mockPush = jest.fn()
 
 const VERSIONS = [
-  { name: 'latest', label: 'latest' },
-  { name: 'v0.5.1', label: 'v0.5.1' },
-  { name: 'v0.4.0', label: 'v0.4.0' },
+  { name: 'latest', label: 'v0.5.x (latest)' },
+  { name: 'v0.4.x', label: 'v0.4.x' },
+  { name: 'v0.3.x', label: 'v0.3.x' },
 ]
 
 const defaultProps = {
@@ -54,26 +54,39 @@ describe('<VersionSelect />', () => {
     expect(options).toHaveLength(VERSIONS.length)
   })
 
-  it('should display the latest version', () => {
-    const { getByRole } = render(<VersionSelect {...defaultProps} />)
-
-    const button = getByRole('button')
-    expect(button).toHaveTextContent('latest')
-  })
-
-  it('should display version from the URL', () => {
+  it('should display the correct version from the URL', () => {
+    // initial, latest
     useRouterMock.mockImplementation(() => {
       return ({
-        route: '/docs/[[...page]]',
-        pathname: '/docs/[[...page]]',
-        asPath: '/docs/v0.5.1/some/nested/article',
+        asPath: '/commands',
         push: mockPush,
       } as unknown) as Router
     })
-    const { getByRole } = render(<VersionSelect {...defaultProps} />)
 
-    const button = getByRole('button')
-    expect(button).toHaveTextContent('v0.5.1')
+    const { getByRole, rerender } = render(<VersionSelect {...defaultProps} />)
+    expect(getByRole('button')).toHaveTextContent('v0.5.x (latest)')
+
+    // simulate navigation to v0.4.x
+    useRouterMock.mockImplementation(() => {
+      return ({
+        asPath: '/commands/v0.4.x',
+        push: mockPush,
+      } as unknown) as Router
+    })
+
+    rerender(<VersionSelect {...defaultProps} />)
+    expect(getByRole('button')).toHaveTextContent('v0.4.x')
+
+    // simulate browser-back to latest
+    useRouterMock.mockImplementation(() => {
+      return ({
+        asPath: '/commands/',
+        push: mockPush,
+      } as unknown) as Router
+    })
+
+    rerender(<VersionSelect {...defaultProps} />)
+    expect(getByRole('button')).toHaveTextContent('v0.5.x (latest)')
   })
 
   it('should navigate to a selected version', () => {
@@ -97,7 +110,7 @@ describe('<VersionSelect />', () => {
     const options = getAllByRole('option')
     userEvent.click(options[index])
 
-    expect(mockPush).toHaveBeenNthCalledWith(1, '/commands/v0.5.1')
+    expect(mockPush).toHaveBeenNthCalledWith(1, '/commands/v0.4.x')
   })
 
   it('should navigate to the latest version', () => {
@@ -105,7 +118,7 @@ describe('<VersionSelect />', () => {
       return ({
         route: '/commands/[[...page]]',
         pathname: '/commands/[[...page]]',
-        asPath: '/commands/v0.5.1',
+        asPath: '/commands/v0.4.x',
         push: mockPush,
       } as unknown) as Router
     })
@@ -122,6 +135,30 @@ describe('<VersionSelect />', () => {
     userEvent.click(options[index])
 
     expect(mockPush).toHaveBeenNthCalledWith(1, '/commands')
+  })
+
+  it('should navigate from a version to another version', () => {
+    useRouterMock.mockImplementation(() => {
+      return ({
+        route: '/commands/[[...page]]',
+        pathname: '/commands/[[...page]]',
+        asPath: '/commands/v0.4.x',
+        push: mockPush,
+      } as unknown) as Router
+    })
+
+    const { getByRole, getAllByRole } = render(
+      <VersionSelect {...defaultProps} />
+    )
+    const index = 2
+
+    const combobox = getByRole('combobox')
+    userEvent.click(combobox)
+
+    const options = getAllByRole('option')
+    userEvent.click(options[index])
+
+    expect(mockPush).toHaveBeenNthCalledWith(1, '/commands/v0.3.x')
   })
 
   it('should navigate to a selected version while retaining the sub path', () => {
@@ -150,7 +187,7 @@ describe('<VersionSelect />', () => {
 
     expect(mockPush).toHaveBeenNthCalledWith(
       1,
-      '/docs/v0.4.0/some/nested/article'
+      '/docs/v0.3.x/some/nested/article'
     )
   })
 
@@ -160,9 +197,9 @@ describe('<VersionSelect />', () => {
         route: '/docs/[[...page]]',
         pathname: '/docs/[[...page]]',
         query: {
-          page: ['v0.4.0', 'some', 'nested', 'article'],
+          page: ['v0.4.x', 'some', 'nested', 'article'],
         },
-        asPath: '/docs/v0.4.0/some/nested/article',
+        asPath: '/docs/v0.4.x/some/nested/article',
         push: mockPush,
       } as unknown) as Router
     })
@@ -179,5 +216,91 @@ describe('<VersionSelect />', () => {
     userEvent.click(options[index])
 
     expect(mockPush).toHaveBeenNthCalledWith(1, '/docs/some/nested/article')
+  })
+
+  it('should navigate from a version to another version while retaining the sub path', () => {
+    useRouterMock.mockImplementation(() => {
+      return ({
+        route: '/docs/[[...page]]',
+        pathname: '/docs/[[...page]]',
+        query: {
+          page: ['v0.4.x', 'some', 'nested', 'article'],
+        },
+        asPath: '/docs/v0.4.x/some/nested/article',
+        push: mockPush,
+      } as unknown) as Router
+    })
+
+    const { getByRole, getAllByRole } = render(
+      <VersionSelect {...defaultProps} />
+    )
+    const index = 2
+
+    const combobox = getByRole('combobox')
+    userEvent.click(combobox)
+
+    const options = getAllByRole('option')
+    userEvent.click(options[index])
+
+    expect(mockPush).toHaveBeenNthCalledWith(
+      1,
+      '/docs/v0.3.x/some/nested/article'
+    )
+  })
+
+  it('should noop if selecting latest, while on latest', () => {
+    useRouterMock.mockImplementation(() => {
+      return ({
+        route: '/docs/[[...page]]',
+        pathname: '/docs/[[...page]]',
+        query: {
+          page: ['some', 'nested', 'article'],
+        },
+        asPath: '/docs/some/nested/article',
+        push: mockPush,
+      } as unknown) as Router
+    })
+
+    const { getByRole, getAllByRole } = render(
+      <VersionSelect {...defaultProps} />
+    )
+    const index = 0
+
+    const combobox = getByRole('combobox')
+    userEvent.click(combobox)
+
+    const options = getAllByRole('option')
+    userEvent.click(options[index])
+
+    expect(mockPush).not.toHaveBeenCalled()
+  })
+
+  it('should noop if selecting a version, while on the same version', () => {
+    const currentVersion = 'v0.4.x'
+
+    useRouterMock.mockImplementation(() => {
+      return ({
+        route: '/docs/[[...page]]',
+        pathname: '/docs/[[...page]]',
+        query: {
+          page: ['some', 'nested', 'article'],
+        },
+        asPath: `/docs/${currentVersion}/some/nested/article`,
+        push: mockPush,
+      } as unknown) as Router
+    })
+
+    const { getByRole, getAllByRole } = render(
+      <VersionSelect {...defaultProps} />
+    )
+    const index = VERSIONS.findIndex((v) => v.name === currentVersion)
+
+    const combobox = getByRole('combobox')
+    userEvent.click(combobox)
+
+    const options = getAllByRole('option')
+    userEvent.click(options[index])
+
+    expect(mockPush).not.toHaveBeenCalled()
   })
 })